***** major change to raw AE addresses, now 32 bit (int) changed from 16 bit (short) throughout jAER. affects logging, playback, socket streams, etc.  may break existing socket connection code, e.g. with blender.
Added classes to support new TCVS320 chip with its new address space and bias generator. Not finished yet.


git-svn-id: https://svn.code.sf.net/p/jaer/code/trunk/host/java@414 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ src/ch/unizh/ini/caviar/aemapper/AbstractAEMapper.java
@@ -61,7 +61,7 @@ abstract public class AbstractAEMapper implements AEMapper {
         if(!isMappingEnabled() || isMappingPassThrough()) return input;
         checkOutputPacket();
         int n=input.getNumEvents();
-        short[] ain=input.getAddresses();
+        int[] ain=input.getAddresses();
         int[] tin=input.getTimestamps();
         outputPacket.setNumEvents(0);
         for(int i=0;i<n;i++){
